task 7up6

+++ 0x00-MySQL_Advanced/7-average_score.sql
@@ -4,11 +4,9 @@
 DELIMITER $$
 
 DROP PROCEDURE IF EXISTS ComputeAverageScoreForUser;
-CREATE PROCEDURE ComputeAverageScoreForUser(
-    IN user_id INT
-)
+CREATE PROCEDURE ComputeAverageScoreForUser(IN user_id INT)
 BEGIN
-    DECLARE avg_score INT;
+    DECLARE avg_score FLOAT;
     SELECT AVG(score) INTO avg_score
     FROM corrections
     WHERE corrections.user_id = user_id;
